feat: 음식점 정보 모달창 즐겨찾기 추가/삭제 기능 구현
- refactor: 파일 분리
- rename: 폴더 및 함수명 변경
