+++ __tests__/render.test.js
@@ -5,7 +5,7 @@
 import { screen } from "@testing-library/dom";
 import "@testing-library/jest-dom";
 
-import { renderRestaurantList } from "../src/component/restaurantList";
+import { renderRestaurants } from "../src/component/restaurants";
 import { initialRestaurantData } from "../src/constant/initialRestaurants";
 
 beforeEach(() => {
@@ -13,7 +13,7 @@ beforeEach(() => {
 });
 
 test("초기 음식점 목록을 렌더링한다.", () => {
-  const restaurants = renderRestaurantList(initialRestaurantData);
+  const restaurants = renderRestaurants(initialRestaurantData);
   document.body.insertAdjacentHTML("beforeend", restaurants);
 
   expect(screen.queryByText("친친")).toBeInTheDocument();
